Update from Google.

+++ src/main/java/com/google/devtools/build/lib/standalone/LocalSpawnStrategy.java
@@ -24,6 +24,7 @@ import com.google.devtools.build.lib.shell.Command;
 import com.google.devtools.build.lib.shell.CommandException;
 import com.google.devtools.build.lib.syntax.Label;
 import com.google.devtools.build.lib.util.CommandFailureUtils;
+import com.google.devtools.build.lib.util.OsUtils;
 import com.google.devtools.build.lib.util.io.FileOutErr;
 import com.google.devtools.build.lib.vfs.Path;
 
@@ -37,12 +38,13 @@ import java.util.List;
 @ExecutionStrategy(name = { "standalone" }, contextType = SpawnActionContext.class)
 public class LocalSpawnStrategy implements SpawnActionContext {
   private final boolean verboseFailures;
-  
+
   private final Path processWrapper;
 
   public LocalSpawnStrategy(Path execRoot, boolean verboseFailures) {
     this.verboseFailures = verboseFailures;
-    this.processWrapper = execRoot.getRelative("_bin/process-wrapper");
+    this.processWrapper = execRoot.getRelative(
+        "_bin/process-wrapper" + OsUtils.executableExtension());
   }
 
   /**
@@ -59,17 +61,23 @@ public class LocalSpawnStrategy implements SpawnActionContext {
     }
 
     // We must wrap the subprocess with process-wrapper to kill the process tree.
-    // All actions therefore depend on the process-wrapper file. Since it's embedded, 
+    // All actions therefore depend on the process-wrapper file. Since it's embedded,
     // we don't bother with declaring it as an input.
     List<String> args = new ArrayList<>();
-    args.add(processWrapper.getPathString());
-    args.add("-1"); /* timeout */
-    args.add("0");  /* kill delay. */
+    if (!OsUtils.isWindows()) {
+      // TODO(bazel-team): process-wrapper seems to work on Windows, but requires
+      // additional setup as it is an msys2 binary, so it needs msys2 DLLs on %PATH%.
+      // Disable it for now to make the setup easier and to avoid further PATH hacks.
+      // Ideally we should have a native implementation of process-wrapper for Windows.
+      args.add(processWrapper.getPathString());
+      args.add("-1"); /* timeout */
+      args.add("0");  /* kill delay. */
 
-    // TODO(bazel-team): use process-wrapper redirection so we don't have to
-    // pass test logs through the Java heap.
-    args.add("-");  /* stdout. */
-    args.add("-");  /* stderr. */
+      // TODO(bazel-team): use process-wrapper redirection so we don't have to
+      // pass test logs through the Java heap.
+      args.add("-");  /* stdout. */
+      args.add("-");  /* stderr. */
+    }
     args.addAll(spawn.getArguments());
 
     String cwd = executor.getExecRoot().getPathString();
